More updates for Firefox & more efficient cursor takeover
Additions for Mozilla keyframes keeps getting deleted, hopefully this enables them.

Also changed the showing and hiding of the cursor to a class to allow for the transform attributes within the style change on the mousemove event to be done as a setAttribute rather than a property change (Firefox didn’t like the self.cursor.style['-moz-transform'] = '…'; method).

A future update will likely be using the classList javascript method rather than className.

Signed-off-by: Matthew Potter <mdpotter@gmail.com>